feat: Introduce new account management models and update existing structures
- Added AccountInstallment, AccountItem, AccountItemSerialItem, AccountTransaction, CreditAccount, LayawayAccount, EmploymentStatus, and EmployeeIdentityUser models.
- Removed AccountPayment model as part of restructuring.
- Updated AccountStatus and AccountType models to include new fields and indices.
- Modified ApplicationDbContext to include new DbSet properties for the newly added models.
- Adjusted relationships and foreign keys across various models to ensure data integrity.
- Updated existing models (e.g., Client, File, Product) to reflect changes in data types and relationships.
- Removed obsolete references to Stock in multiple models to streamline the database schema.

+++ SemaforoSystem.Server/Models/File.cs
@@ -26,7 +26,7 @@ public partial class File
     public int? SchoolId { get; set; }
 
     [Column("account_id")]
-    public int? AccountId { get; set; }
+    public long? AccountId { get; set; }
 
     [Column("provider_account_id")]
     public int? ProviderAccountId { get; set; }
